Refactoring to a more modular structure
I want this to feel more like a toolbox with a variety
of different tools to use in different situations.

For example, by creating an abstraction layer for working
with servers, we can eventually replace the need for something
like pssh with a command like:

Server.all.run('some command -on all servers')

We can also have convenience methods for working with individual
servers:

Server.find('cm').service('ntp').restart

Or methods for manipulating groups of servers:

Server.where(role: 'worker')

And if we can track all of this information in a config file,
we can have the config file be the single source of truth, from
which we can build and configure an entire cluster.